Edit rename all pascal case folder names to camel case
