+++ src/App.tsx
@@ -1,38 +1,38 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import Home from '@pages/User/Home';
-import AdminHome from '@pages/Admin/AdminHome';
-import Login from '@pages/User/Login';
-import AdminWorkspace from '@pages/Admin/AdminWorkspace';
-import Register from '@pages/User/Register';
-import AdminProduct from '@pages/Admin/AdminProduct';
-import Order from '@pages/User/order/Order';
-import AdminAccount from '@pages/Admin/AdminAccount';
+import Home from '@pages/user/Home';
+import AdminHome from '@pages/admin/AdminHome';
+import Login from '@pages/user/Login';
+import AdminWorkspace from '@pages/admin/AdminWorkspace';
+import Register from '@pages/user/Register';
+import AdminProduct from '@pages/admin/AdminProduct';
+import Order from '@pages/user/order/Order';
+import AdminAccount from '@pages/admin/AdminAccount';
 import { RecoilRoot } from 'recoil';
 import LoadingModal from '@components/common/modal/LoadingModal';
 import { Global } from '@emotion/react';
 import { globalStyles } from '@styles/globalStyles';
-import OrderBasket from '@pages/User/order/OrderBasket';
-import OrderPay from '@pages/User/order/OrderPay';
-import AdminProductAdd from '@pages/Admin/AdminProductAdd';
-import OrderComplete from '@pages/User/order/OrderComplete';
-import AdminOrderHistory from '@pages/Admin/order/AdminOrderHistory';
-import AdminProductEdit from '@pages/Admin/AdminProductEdit';
+import OrderBasket from '@pages/user/order/OrderBasket';
+import OrderPay from '@pages/user/order/OrderPay';
+import AdminProductAdd from '@pages/admin/AdminProductAdd';
+import OrderComplete from '@pages/user/order/OrderComplete';
+import AdminOrderHistory from '@pages/admin/order/AdminOrderHistory';
+import AdminProductEdit from '@pages/admin/AdminProductEdit';
 import ReactGA from 'react-ga4';
 import RouterChangeTracker from './RouterChangeTracker';
-import AdminMyInfo from '@pages/Admin/AdminMyInfo';
-import AdminProductCategories from '@pages/Admin/AdminProductCategoires';
-import AdminTableCount from '@pages/Admin/order/AdminTableCount';
-import ResetPassword from '@pages/User/ResetPassword';
-import Info from '@pages/User/Info';
-import SuperAdminHome from '@pages/SuperAdmin/SuperAdminHome';
-import SuperAdminWorkspace from '@pages/SuperAdmin/SuperAdminWorkspace';
-import SuperAdminManage from '@pages/SuperAdmin/SuperAdminManage';
-import SuperAdminUser from '@pages/SuperAdmin/SuperAdminUser';
-import AdminOrder from '@pages/Admin/order/AdminOrder';
-import AdminOrderTable from '@pages/Admin/order/AdminOrderTable';
-import AdminOrderTableHistory from '@pages/Admin/order/AdminOrderTableHistory';
-import AdminOrderRealtime from '@pages/Admin/order/AdminOrderRealtime';
+import AdminMyInfo from '@pages/admin/AdminMyInfo';
+import AdminProductCategories from '@pages/admin/AdminProductCategoires';
+import AdminTableCount from '@pages/admin/order/AdminTableCount';
+import ResetPassword from '@pages/user/ResetPassword';
+import Info from '@pages/user/Info';
+import SuperAdminHome from '@pages/super-admin/SuperAdminHome';
+import SuperAdminWorkspace from '@pages/super-admin/SuperAdminWorkspace';
+import SuperAdminManage from '@pages/super-admin/SuperAdminManage';
+import SuperAdminUser from '@pages/super-admin/SuperAdminUser';
+import AdminOrder from '@pages/admin/order/AdminOrder';
+import AdminOrderTable from '@pages/admin/order/AdminOrderTable';
+import AdminOrderTableHistory from '@pages/admin/order/AdminOrderTableHistory';
+import AdminOrderRealtime from '@pages/admin/order/AdminOrderRealtime';
 
 ReactGA.initialize('G-XGYLSPGK2G');
 function App() {
