+++ src/pages/admin/AdminProduct.tsx
