+++ src/pages/user/Info.tsx
